refactor: rename `realtime.d.ts` => `realtime.ts`
I'd like to add an `enum` to `realtime.d.ts` but turns out you can't really add
and `enum` to a "definitions" file
