+++ assets/src/components/notificationDrawer.tsx
@@ -3,7 +3,7 @@ import { NotificationsContext } from "../contexts/notificationsContext"
 import { StateDispatchContext } from "../contexts/stateDispatchContext"
 import useCurrentTime from "../hooks/useCurrentTime"
 import { markAllAsRead } from "../hooks/useNotificationsReducer"
-import { Notification } from "../realtime.d"
+import { Notification } from "../realtime"
 import { rememberNotificationDrawerScrollPosition } from "../state"
 import { openVPPForNotification } from "./notifications"
 import { NotificationCard } from "./notificationCard"
