+++ assets/src/components/notificationModals/inactiveNotificationModal.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import Loading from "../loading"
 import { useMinischeduleRuns } from "../../hooks/useMinischedule"
 import { Activity, Run, Piece, Trip } from "../../minischedule"
-import { Notification, RunId } from "../../realtime.d"
+import { Notification, RunId } from "../../realtime"
 import { now, serviceDaySeconds } from "../../util/dateTime"
 import { title } from "../notificationCard"
 import BasicNotificationModal from "./basicNotificationModal"
