+++ assets/src/components/propertiesPanel.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react"
 import useSocket from "../hooks/useSocket"
 import useVehicleForId from "../hooks/useVehicleForId"
 import { isLoggedOut, isVehicle } from "../models/vehicle"
-import { Ghost, Vehicle } from "../realtime.d"
+import { Ghost, Vehicle } from "../realtime"
 import GhostPropertiesPanel from "./propertiesPanel/ghostPropertiesPanel"
 import StaleDataPropertiesPanel from "./propertiesPanel/staleDataPropertiesPanel"
 import VehiclePropertiesPanel from "./propertiesPanel/vehiclePropertiesPanel"
