+++ assets/src/components/vehicleIcon.tsx
@@ -6,7 +6,7 @@ import { DrawnStatus, statusClasses } from "../models/vehicleStatus"
 import { AlertIconStyle, IconAlertCircleSvgNode } from "./iconAlertCircle"
 import { runIdToLabel } from "../helpers/vehicleLabel"
 import { isGhost } from "../models/vehicle"
-import { Ghost, RunId, Vehicle, VehicleInScheduledService } from "../realtime.d"
+import { Ghost, RunId, Vehicle, VehicleInScheduledService } from "../realtime"
 import { BlockId, ViaVariant } from "../schedule.d"
 import { scheduleAdherenceLabelString } from "./propertiesPanel/header"
 import { UserSettings } from "../userSettings"
