+++ assets/tests/hooks/useShuttleVehicles.test.ts
@@ -1,7 +1,7 @@
 import { describe, test, expect } from "@jest/globals"
 import { renderHook } from "@testing-library/react"
 import useShuttleVehicles from "../../src/hooks/useShuttleVehicles"
-import { Vehicle } from "../../src/realtime.d"
+import { Vehicle } from "../../src/realtime"
 import { makeMockChannel, makeMockSocket } from "../testHelpers/socketHelpers"
 import { vehicleDataFactory } from "../factories/vehicle_data"
 import { vehicleFromData } from "../../src/models/vehicleData"
