+++ assets/tests/hooks/useVehicles.test.ts
@@ -3,7 +3,7 @@ import { renderHook } from "@testing-library/react"
 import useVehicles from "../../src/hooks/useVehicles"
 import * as browser from "../../src/models/browser"
 import { VehicleData, vehicleFromData } from "../../src/models/vehicleData"
-import { Ghost, Vehicle } from "../../src/realtime.d"
+import { Ghost, Vehicle } from "../../src/realtime"
 import { RouteId } from "../../src/schedule.d"
 import { makeMockChannel, makeMockSocket } from "../testHelpers/socketHelpers"
 import ghostFactory from "../factories/ghost"
