+++ assets/tsconfig.json
@@ -4,7 +4,6 @@
   "files": [
     "src/skate.d.ts",
     "src/schedule.d.ts",
-    "src/realtime.d.ts",
     "src/minischedule.d.ts"
   ],
   "compilerOptions": {
